We no longer use BITMAP_FORMAT_RGB15 in MESS (no whatsnew)
git-svn-id: svn://messdev.no-ip.org/mess@13847 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/includes/svision.h
@@ -23,7 +23,7 @@ typedef struct
 
 typedef struct
 {
-	UINT16 palette[4/*0x40?*/]; /* rgb8 */
+	UINT32 palette[4/*0x40?*/]; /* rgb8 */
 	int palette_on;
 } tvlink_t;
